document 'Physics' group, related to external-simulator physics def > decls/def sound/*.sndshd > decls/sound/*.sndshd efx/ > decls/efx NSWeapon: add "snd_fireStart", "snd_fireStop", "snd_fireLoop" NSWeapon: add "view_geomset" to FireInfo NSWeapon: add "knockbackRelease" to FireInfo NSWeapon: add "altMode" toggle NSWeapon: "actLoop" decoupled from "actFireStart" NSWeapon: fireRate inherited from model if not set NSWeapon: query all sequences from FireInfo NSWeapon privatize weapon state solidify more def documentation across the board NSWeapon add "chargeTime" key rework state machine in NSWeapon reload timing separation in NSWeapon add overheatLength, overheatPerShot to NSWeapon

+++ src/shared/sound.h
@@ -216,7 +216,7 @@ void Sound_Speak(entity targetEntity, string sentencesEntry);
 #endif
 
 /** Stops sounds on a given channel, on a target entity. */
-void Sound_Stop(entity target, int chan);
+void Sound_Stop(entity target, float chan);
 
 void Sound_DebugList();
